// star-wars/src/interfaces/Character.ts
export interface CharacterProperties {
    birth_year: string;
    created: string;
    edited: string;
    eye_color: string;
    gender: string;
    hair_color: string;
    height: string;
    homeworld: string;
    mass: string;
    name: string;
    skin_color: string;
    url: string;
  }
  
 export interface Character {
    uid:string,
    properties:CharacterProperties,
    description: string;
    __v: number;         
    _id: string; 
  }